Вынесла код из одного файла в модули
Добавила файлы board.js, cell.js. Настроила работу импортов.
