+++ style.css
@@ -83,6 +83,7 @@ header > div {
 	font-size: 40px;
     font-weight: bold;
 }
+
 /* add left margin to first childs of rows */
 .cell:nth-child(4n+1) {
     margin-left: 2vmin;
