--- text.txt
@@ -1,4 +1,0 @@
-1) Генерируем два числа 2 для начала игры в случайных клетках.
-2) На события клавиш вверх/вниз/вправо/влево перемещаем все числа в соответствующую сторону и если цифры рядом стоящие одинаковые складываем,
- если разные то смещаем число до крайнего заполненого поля.
-3) При нажатии на клавишу перемещения генерируется в случайной клетке цифра 2.
